requirements.txt
# requirements.txt

# Core Langchain & LangGraph (Usa las versiones más recientes o las que te funcionaron)
langchain>=0.2.0 
langgraph>=0.1.0 
langchain-core>=0.2.0 
langchain-anthropic>=0.1.0 

# Langchain Community (Necesaria para algunas utilidades base)
langchain-community>=0.2.0

# Cálculo Numérico
numpy>=1.20 
numpy-financial==1.0.0 # <--- Versión corregida a la última estable conocida
scipy>=1.7 # Para norm.cdf en Black-Scholes

# Validación y Modelos
pydantic>=2.0

# Interfaz Web
streamlit>=1.30 

# Para cargar .env localmente (Recomendado)
python-dotenv>=1.0.0